model/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.21.0
